interoperability_adapter/umls_etl/requirements.txt
psycopg2-binary~=2.9.7
